+++ en/tools/quicklook/Summary.aspx.cs
@@ -26,7 +26,7 @@ namespace SkyServer.Tools.QuickLook
 
          long? id = null;
          string apid;
-         long? specId = null;
+         decimal? specId = null;
          string sidstring = null;
          double? qra = null;
          double? qdec = null;
